+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog
 =========
 
+## 1.0.0 (2020-01-05)
+ * Added support for Pages and News models
+ * Added console, migrations and controller
+ 
 ## 0.0.1 (2019-12-20)
  * Initial commit